refactor(circuit): extract LifecycleMixin from CircuitRoutines [Clean Code: SRP]
Tier 5b-B. Move the 16-method "parameter sync + dispatch handling +
watched-property installation" cluster from
``circuit_internals/routines.py`` into a new sibling mixin
``circuit_internals/lifecycle.py`` (528 lines).  CircuitRoutines now
subclasses three sibling mixins:

    class CircuitRoutines(
        LifecycleMixin,
        SubsystemTreeMixin,
        HamiltonianAssemblyMixin,
        ABC,
    ):

The new mixin owns:

* parameter syncing across the subsystem tree
  (``_sync_parameters_with_parent``,
  ``_sync_parameters_with_subsystems``,
  ``_set_sync_status_to_True``);
* central-dispatch event reception
  (``receive``, ``_store_updated_subsystem_index``,
  ``_mark_all_subsystems_as_affected``,
  ``_propagate_param_to_affected_subsystems``);
* the update pipeline
  (``update``, ``_perform_internal_updates``, ``_update_bare_esys``,
  ``_is_internal_update_required``, ``_fetch_symbolic_hamiltonian``);
* watched-property setter installation
  (``_make_property``, the three
  ``_set_property_and_update_<param_kind>`` methods).

The module-level helpers consumed exclusively by this cluster move
with it: the ``_dispatch_suspended`` context manager, the
``PropertyUpdateType`` Literal alias, and the
``_PROPERTY_SETTER_BY_TYPE`` lookup.

The MRO order (``LifecycleMixin``, ``SubsystemTreeMixin``,
``HamiltonianAssemblyMixin``, ``ABC``) preserves attribute-lookup
behaviour: every method in every mixin still resolves through
``self`` exactly as before.  Cross-mixin attributes / callables
declared under ``TYPE_CHECKING`` so mypy resolves
``self.<x>`` correctly.

routines.py: 1041 -> 655 lines (-37%).  Cumulative effect of the
three Tier 5 extractions: 2447 -> 655 lines, a 73% reduction.  The
residual file now holds only:

* serialization (``serialize`` / ``deserialize``),
* Hilbert-space basics (``cutoffs_dict``,
  ``discretized_grids_dict_for_vars``, ``hilbertdim``,
  ``_kron_operator``, ``_sparsity_adaptive``, ``_identity_qobj``,
  ``_identity``, ``get_cutoffs``, ``_collect_cutoff_values``),
* bare-eigensystem caching (``generate_bare_eigensys``,
  ``set_bare_eigensys``),
* a handful of accessor / setter helpers
  (``set_discretized_phi_range``, ``set_and_return``,
  ``get_ext_basis``, ``get_osc_param``, ``_get_cutoff_value``,
  ``return_root_child``, ``return_parent_circuit``,
  ``default_params``, ``__setattr__``, ``__reduce__``,
  ``__setstate__``, ``create``).

Verified against the characterization-test goldens
(``rtol=1e-10`` on the four ``hamiltonian()`` / ``eigenvals()``
fixtures) plus the five lifecycle-dispatch tests, all passing.
Full pytest: 322 passed, 14 skipped.  mypy clean (103 source
files, +1 for the new module).
